Always cast ntohl to uint32_t and ntohs to uint16_t

+++ worker/src/RTC/RTCP/Feedback.cpp
@@ -84,8 +84,8 @@ namespace RTC { namespace RTCP
 	{
 		MS_TRACE();
 
-		MS_DEBUG_DEV("  sender ssrc : %" PRIu32, ntohl(this->header->s_ssrc));
-		MS_DEBUG_DEV("  media ssrc  : %" PRIu32, ntohl(this->header->m_ssrc));
+		MS_DEBUG_DEV("  sender ssrc : %" PRIu32, (uint32_t)ntohl(this->header->s_ssrc));
+		MS_DEBUG_DEV("  media ssrc  : %" PRIu32, (uint32_t)ntohl(this->header->m_ssrc));
 		MS_DEBUG_DEV("  size        : %zu", this->GetSize());
 	}
