+++ worker/src/RTC/RTCP/FeedbackPsFir.cpp
@@ -57,7 +57,7 @@ namespace RTC { namespace RTCP
 		MS_TRACE();
 
 		MS_DEBUG_DEV("<FirItem>");
-		MS_DEBUG_DEV("  ssrc            : %" PRIu32, ntohl(this->header->ssrc));
+		MS_DEBUG_DEV("  ssrc            : %" PRIu32, (uint32_t)ntohl(this->header->ssrc));
 		MS_DEBUG_DEV("  sequence number : %" PRIu8, this->header->sequence_number);
 		MS_DEBUG_DEV("\t\t</FirItem>");
 	}
