+++ worker/src/RTC/RTCP/FeedbackPsLei.cpp
@@ -49,7 +49,7 @@ namespace RTC { namespace RTCP
 		MS_TRACE();
 
 		MS_DEBUG_DEV("<PsLeiItem>");
-		MS_DEBUG_DEV("  ssrc : %" PRIu32, ntohl(this->header->ssrc));
+		MS_DEBUG_DEV("  ssrc : %" PRIu32, (uint32_t)ntohl(this->header->ssrc));
 		MS_DEBUG_DEV("</PsLeiItem>");
 	}
 }}
