+++ worker/src/RTC/RTCP/FeedbackRtpEcn.cpp
@@ -41,13 +41,13 @@ namespace RTC { namespace RTCP
 		MS_TRACE();
 
 		MS_DEBUG_DEV("<EcnItem>");
-		MS_DEBUG_DEV("  sequence number    : %" PRIu32, ntohl(this->header->sequence_number));
-		MS_DEBUG_DEV("  ect0 counter       : %" PRIu32, ntohl(this->header->ect0_counter));
-		MS_DEBUG_DEV("  ect1 counter       : %" PRIu32, ntohl(this->header->ect1_counter));
-		MS_DEBUG_DEV("  ecn ce counter     : %" PRIu16, ntohs(this->header->ecn_ce_counter));
-		MS_DEBUG_DEV("  not ect counter    : %" PRIu16, ntohs(this->header->not_ect_counter));
-		MS_DEBUG_DEV("  lost packets       : %" PRIu16, ntohs(this->header->lost_packets));
-		MS_DEBUG_DEV("  duplicated packets : %" PRIu16, ntohs(this->header->duplicated_packets));
+		MS_DEBUG_DEV("  sequence number    : %" PRIu32, (uint32_t)ntohl(this->header->sequence_number));
+		MS_DEBUG_DEV("  ect0 counter       : %" PRIu32, (uint32_t)ntohl(this->header->ect0_counter));
+		MS_DEBUG_DEV("  ect1 counter       : %" PRIu32, (uint32_t)ntohl(this->header->ect1_counter));
+		MS_DEBUG_DEV("  ecn ce counter     : %" PRIu16, (uint16_t)ntohs(this->header->ecn_ce_counter));
+		MS_DEBUG_DEV("  not ect counter    : %" PRIu16, (uint16_t)ntohs(this->header->not_ect_counter));
+		MS_DEBUG_DEV("  lost packets       : %" PRIu16, (uint16_t)ntohs(this->header->lost_packets));
+		MS_DEBUG_DEV("  duplicated packets : %" PRIu16, (uint16_t)ntohs(this->header->duplicated_packets));
 		MS_DEBUG_DEV("</EcnItem>");
 	}
 }}
